# contracts/catalog/openapi.json
{
  "components": {
    "schemas": {
...
      "BroadcastResponse": {
        "description": "Response model for test broadcast operations",
        "properties": {
          "active_connections": {
            "description": "Number of active connections that received the event",
            "example": 3,
            "title": "Active Connections",
            "type": "integer"
          },
          "message": {
            "description": "Descriptive message about the operation",
            "example": "Test event broadcasted to devices channel",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "description": "Operation status",
            "example": "success",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message",
          "active_connections"
        ],
        "title": "BroadcastResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SSEStats": {
        "description": "SSE connection statistics response model",
        "properties": {
          "active_connections": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Number of active connections per channel",
            "example": {
              "devices": 3,
              "scenarios": 1,
              "system": 2
            },
            "title": "Active Connections",
            "type": "object"
          },
          "channels": {
            "description": "List of available SSE channels",
            "example": [
              "devices",
              "scenarios",
              "system"
            ],
            "items": {
              "type": "string"
            },
            "title": "Channels",
            "type": "array"
          },
          "total_connections": {
            "description": "Total number of active connections across all channels",
            "example": 6,
            "title": "Total Connections",
            "type": "integer"
          }
        },
        "required": [
          "active_connections",
          "total_connections",
          "channels"
        ],
        "title": "SSEStats",
        "type": "object"
      },
...
      "TestEventData": {
        "description": "Test event data model for broadcasting",
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional additional data",
            "example": {
              "key": "value",
              "number": 42
            },
            "title": "Data"
          },
          "message": {
            "description": "Test message content",
            "example": "This is a test event",
            "title": "Message",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional timestamp for the event",
            "example": "2024-01-01T12:00:00Z",
            "title": "Timestamp"
          }
        },
        "required": [
          "message"
        ],
        "title": "TestEventData",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A web service that manages MQTT devices with typed configurations",
    "title": "MQTT Web Service",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/events/devices": {
      "get": {
        "description": "Subscribe to real-time device events via Server-Sent Events",
        "operationId": "devices_stream_events_devices_get",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "SSE service not initialized"
                }
              }
            },
            "description": "Service Unavailable - SSE service not initialized"
          }
        },
        "summary": "Device Events Stream",
        "tags": [
          "Server-Sent Events"
        ]
      }
    },
    "/events/scenarios": {
      "get": {
        "description": "Subscribe to real-time scenario events via Server-Sent Events",
        "operationId": "scenarios_stream_events_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "SSE service not initialized"
                }
              }
            },
            "description": "Service Unavailable - SSE service not initialized"
          }
        },
        "summary": "Scenario Events Stream",
        "tags": [
          "Server-Sent Events"
        ]
      }
    },
    "/events/stats": {
      "get": {
        "description": "Get current statistics about active Server-Sent Events connections",
        "operationId": "get_sse_stats_events_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "active_connections": {
                    "devices": 3,
                    "scenarios": 1,
                    "system": 2
                  },
                  "channels": [
                    "devices",
                    "scenarios",
                    "system"
                  ],
                  "total_connections": 6
                },
                "schema": {
                  "$ref": "#/components/schemas/SSEStats"
                }
              }
            },
            "description": "Connection statistics retrieved successfully"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "SSE service not initialized"
                }
              }
            },
            "description": "Service Unavailable - SSE service not initialized"
          }
        },
        "summary": "SSE Connection Statistics",
        "tags": [
          "Server-Sent Events"
        ]
      }
    },
    "/events/system": {
      "get": {
        "description": "Subscribe to real-time system-wide events via Server-Sent Events",
        "operationId": "system_stream_events_system_get",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "SSE service not initialized"
                }
              }
            },
            "description": "Service Unavailable - SSE service not initialized"
          }
        },
        "summary": "System Events Stream",
        "tags": [
          "Server-Sent Events"
        ]
      }
    },
    "/events/test/{channel}": {
      "post": {
        "description": "Send a test event to a specific SSE channel for development and testing",
        "operationId": "test_broadcast_events_test__channel__post",
        "parameters": [
          {
            "description": "SSE channel to broadcast to",
            "example": "devices",
            "in": "path",
            "name": "channel",
            "required": true,
            "schema": {
              "description": "SSE channel to broadcast to",
              "pattern": "^(devices|scenarios|system)$",
              "title": "Channel",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "data": {
                  "sequence": 1,
                  "test": true
                },
                "message": "This is a test event",
                "timestamp": "2024-01-01T12:00:00Z"
              },
              "schema": {
                "$ref": "#/components/schemas/TestEventData",
                "description": "Test event data to broadcast to connected clients"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "active_connections": 3,
                  "message": "Test event broadcasted to devices channel",
                  "status": "success"
                },
                "schema": {
                  "$ref": "#/components/schemas/BroadcastResponse"
                }
              }
            },
            "description": "Test event broadcasted successfully"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid channel: invalid_channel"
                }
              }
            },
            "description": "Invalid channel name provided"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "SSE service not initialized"
                }
              }
            },
            "description": "Service Unavailable - SSE service not initialized"
          }
        },
        "summary": "Broadcast Test Event",
        "tags": [
          "Server-Sent Events"
        ]
      }
    },
...
  }
}